+++ src/api-server/router/index.ts
@@ -1,9 +1,11 @@
 import { Router } from 'express';
 
 import forumRoutes from './forumRoutes';
+import feedbackRoutes from './feedbackRoutes';
 
 const router: Router = Router();
 
 forumRoutes(router);
+feedbackRoutes(router);
 
 export default router;
