--- src/feedback-server/router/index.ts
@@ -1,9 +1,0 @@
-import { Router } from 'express';
-
-import feedbackRoutes from './feedbackRoutes';
-
-const router: Router = Router();
-
-feedbackRoutes(router);
-
-export default router;
