+++ webpack.config.ts
@@ -1,11 +1,9 @@
 import clientConfig from './webpack/client.config';
 import serverConfig from './webpack/server.config';
 import apiServerConfig from './webpack/api-server.config';
-import feedbackServerConfig from './webpack/feedback-server.config';
 
 export default [
   clientConfig,
   serverConfig,
   apiServerConfig,
-  feedbackServerConfig,
 ];
